arm64: dts: report SW_LID events for the main hall sensor across all devices
Android expects SW_LID events to be reported on lid/flip cover opening
and closing. Samsung has its own custom SW_FLIP event that is reported
instead and unrecognized.

Interestingly, the raw value for the event to report for the hall sensor
is set in device trees. The value for SW_FLIP, 0x15, is set in all of
them. Set the value for SW_LID, 0x00.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>
Signed-off-by: eun0115 <gianpaoloestacio5@gmail.com>